Represent returning with optional value instead of mutable state.
Previously, we used a mutable state in the environment to represent whether the interpreter was returning.  Because only the return statement should return a value, we use an Option<Value> in execute to represent that a return is currently propegating.  This should be simpler to handle, and we can switch the Option out for another enum later if we want to represent something like exceptions as well.